# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LLM Eiskunstlauf Championship API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gateways:
    get:
      operationId: listGateways
      tags: [gateways]
      summary: List all configured LLM gateways
      responses:
        "200":
          description: List of gateways
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Gateway"
    post:
      operationId: createGateway
      tags: [gateways]
      summary: Add a new LLM gateway
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGateway"
      responses:
        "201":
          description: Gateway created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gateway"
  /gateways/{id}:
    delete:
      operationId: deleteGateway
      tags: [gateways]
      summary: Delete a gateway
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /gateways/{id}/models:
    get:
      operationId: listGatewayModels
      tags: [models]
      summary: List available models from a specific gateway
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModelInfo"
  /datasets:
    get:
      operationId: listDatasets
      tags: [datasets]
      summary: List all datasets
      responses:
        "200":
          description: List of datasets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dataset"
    post:
      operationId: createDataset
      tags: [datasets]
      summary: Create a new dataset (upload markdown)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDataset"
      responses:
        "201":
          description: Dataset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
  /datasets/upload:
    post:
      operationId: uploadDataset
      tags: [datasets]
      summary: Upload a markdown file as a new dataset
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: A .md (Markdown) file
                name:
                  type: string
                  description: Dataset name (defaults to filename if omitted)
                systemPrompt:
                  type: string
                  description: System prompt for evaluating responses
              required:
                - file
                - systemPrompt
      responses:
        "201":
          description: Dataset created from uploaded file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
        "400":
          description: Invalid file type or content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /datasets/{id}:
    get:
      operationId: getDataset
      tags: [datasets]
      summary: Get a single dataset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dataset detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
    delete:
      operationId: deleteDataset
      tags: [datasets]
      summary: Delete a dataset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /datasets/{id}/privacy-check:
    post:
      operationId: privacyCheckDataset
      tags: [datasets]
      summary: Run a privacy/PII check on a dataset using an LLM
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrivacyCheckRequest"
      responses:
        "200":
          description: Privacy check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacyCheckResult"
  /datasets/{id}/anonymize:
    post:
      operationId: anonymizeDataset
      tags: [datasets]
      summary: Anonymize PII in a dataset using an LLM
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnonymizeRequest"
      responses:
        "200":
          description: Anonymized dataset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
  /datasets/generate:
    post:
      operationId: generateDataset
      tags: [datasets]
      summary: Generate a test dataset using an LLM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDatasetRequest"
      responses:
        "201":
          description: Generated dataset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
...
components:
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Gateway:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [openrouter, github_copilot, custom]
        baseUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - baseUrl
        - createdAt
    CreateGateway:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [openrouter, github_copilot, custom]
        baseUrl:
          type: string
        apiKey:
          type: string
      required:
        - name
        - type
        - baseUrl
        - apiKey
    ModelInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        gatewayId:
          type: integer
        gatewayName:
          type: string
      required:
        - id
        - name
        - gatewayId
        - gatewayName
    Dataset:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        content:
          type: string
        systemPrompt:
          type: string
        privacyStatus:
          type: string
          enum: [unchecked, clean, issues_found, anonymized]
        privacyReport:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - content
        - systemPrompt
        - privacyStatus
        - createdAt
    CreateDataset:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
        systemPrompt:
          type: string
      required:
        - name
        - content
        - systemPrompt
    PrivacyCheckRequest:
      type: object
      properties:
        gatewayId:
          type: integer
        modelId:
          type: string
      required:
        - gatewayId
        - modelId
    PrivacyCheckResult:
      type: object
      properties:
        status:
          type: string
          enum: [clean, issues_found]
        findings:
          type: array
          items:
            $ref: "#/components/schemas/PrivacyFinding"
        report:
          type: string
      required:
        - status
        - findings
        - report
    PrivacyFinding:
      type: object
      properties:
        type:
          type: string
        text:
          type: string
        suggestion:
          type: string
      required:
        - type
        - text
        - suggestion
    AnonymizeRequest:
      type: object
      properties:
        gatewayId:
          type: integer
        modelId:
          type: string
      required:
        - gatewayId
        - modelId
    GenerateDatasetRequest:
      type: object
      properties:
        name:
          type: string
        topic:
          type: string
        systemPrompt:
          type: string
        numberOfItems:
          type: integer
        gatewayId:
          type: integer
        modelId:
          type: string
      required:
        - name
        - topic
        - systemPrompt
        - numberOfItems
        - gatewayId
        - modelId
...